+++ src/roq/bitmex/json/utils.h
@@ -250,7 +250,7 @@ inline json::ExecInst map(roq::ExecutionInstruction execution_instruction) {
   return json::ExecInst::UNDEFINED;
 }
 
-extern Error guess_error(const std::string_view &message);
+extern roq::Error guess_error(const std::string_view &message);
 
 }  // namespace json
 }  // namespace bitmex