+++ src/roq/bitmex/order_update.cpp
@@ -14,74 +14,30 @@ using namespace roq::literals;
 namespace roq {
 namespace bitmex {
 
-namespace {
-OrderStatus compute_order_status(json::OrdStatus ord_status, bool working_status) {
-  switch (ord_status) {
-    case json::OrdStatus::UNDEFINED:
-    case json::OrdStatus::UNKNOWN:
-      // note! back-stop in case we didn't parse OrdStatus
-      if (working_status)
-        return OrderStatus::WORKING;
-      break;
-    case json::OrdStatus::PENDING_NEW:
-      return OrderStatus::SENT;
-    case json::OrdStatus::NEW:
-    case json::OrdStatus::TRIGGERED:
-      if (!working_status)
-        return OrderStatus::ACCEPTED;
-      else
-        return OrderStatus::WORKING;
-    case json::OrdStatus::DONE_FOR_DAY:
-      return OrderStatus::SUSPENDED;
-    case json::OrdStatus::PARTIALLY_FILLED:
-      return OrderStatus::WORKING;
-    case json::OrdStatus::STOPPED:
-      return OrderStatus::STOPPED;
-    case json::OrdStatus::FILLED:
-      return OrderStatus::COMPLETED;
-    case json::OrdStatus::EXPIRED:
-      return OrderStatus::EXPIRED;
-    case json::OrdStatus::CANCELED:
-      return OrderStatus::CANCELED;
-    case json::OrdStatus::REJECTED:
-      return OrderStatus::REJECTED;
-    case json::OrdStatus::PENDING_CANCEL:  // XXX HANS how to deal with?
-      break;
-    case json::OrdStatus::UNTRIGGERED:  // XXX HANS have no idea what this means...
-      log::warn("Unexpected: ord_status={}, working_status={}"_sv, ord_status, working_status);
-      break;
-  }
-  return {};
-}
-
-RequestStatus compute_request_status(json::OrdStatus ord_status) {
-  if (ord_status == json::OrdStatus::REJECTED)
-    return RequestStatus::REJECTED;
-  return {};
-}
-}  // namespace
+// drop copy
 
 void OrderUpdate::operator()(
     const json::OrderItem &order_item, const server::TraceInfo &trace_info, bool download) {
-  if (!Flags::rest_allow_order_updates())  // XXX HANS do we still need this?
-    return;
   auto status = compute_order_status(order_item.ord_status, order_item.working_indicator);
   log::debug("status={}"_sv, status);
   auto side = json::map(order_item.side);
   auto external_account = fmt::format("{}"_sv, order_item.account);  // XXX alloc
+  auto external_order_id = order_item.order_id;
   auto order_type = json::map(order_item.ord_type);
   auto time_in_force = json::map(order_item.time_in_force);
-  // XXX TODO(thraneh): execution_instruction
-  auto request_status = compute_request_status(order_item.ord_status);
-  auto request_id = order_item.cl_ord_id;
+  auto request_type = order_item.ord_status == json::OrdStatus::CANCELED ? RequestType::CANCEL_ORDER
+                                                                         : RequestType::UNDEFINED;
+  auto request_status = order_item.ord_status == json::OrdStatus::REJECTED
+                            ? RequestStatus::REJECTED
+                            : RequestStatus::ACCEPTED;
   server::OMS_Ack ack{
-      .type = {},
+      .type = request_type,
       .origin = Origin::EXCHANGE,
       .status = request_status,
       .error = json::guess_error(order_item.ord_rej_reason),
       .text = order_item.text,
       .version = {},
-      .request_id = request_id,
+      .request_id = {},  // cancel does not rewrite
       .quantity = order_item.order_qty,
       .price = order_item.price,
   };
@@ -96,12 +52,12 @@ void OrderUpdate::operator()(
       .max_show_quantity = NaN,
       .order_type = order_type,
       .time_in_force = time_in_force,
-      .execution_instruction = {},  // XXX TODO(thraneh)
+      .execution_instruction = {},
       .order_template = {},
       .create_time_utc = {},
       .update_time_utc = order_item.timestamp,  // XXX transact_time?
       .external_account = external_account,
-      .external_order_id = order_item.order_id,
+      .external_order_id = external_order_id,
       .status = status,
       .quantity = order_item.order_qty,
       .price = order_item.price,
@@ -112,20 +68,20 @@ void OrderUpdate::operator()(
       .last_traded_quantity = NaN,
       .last_traded_price = NaN,
       .last_liquidity = {},
-      .routing_id = {},  // XXX TODO(thraneh): decode clOrdID ?
+      .routing_id = {},
       .max_request_version = {},
       .max_response_version = {},
       .max_accepted_version = {},
   };
   if (download) {
-    if (shared_.create_order(request_id, stream_id_, trace_info, order_update)) {
+    if (shared_.create_order(order_item.cl_ord_id, stream_id_, trace_info, order_update)) {
     } else {
       log::warn("*** EXTERNAL ORDER ***"_sv);
       log::warn("order_item={}"_sv, order_item);
     }
   } else {
     if (shared_.update_order(
-            request_id,
+            order_item.cl_ord_id,
             stream_id_,
             trace_info,
             ack,
@@ -143,6 +99,94 @@ void OrderUpdate::operator()(
   log::debug("order={}"_sv, order);
   for (auto &iter : order.data)
     (*this)(iter, trace_info, download);
+}
+
+// order entry
+
+void OrderUpdate::operator()(
+    const json::OrderItem &order_item,
+    const server::TraceInfo &trace_info,
+    RequestType request_type,
+    [[maybe_unused]] uint8_t user_id,
+    uint32_t order_id,
+    uint32_t version) {
+  auto status = compute_order_status(order_item.ord_status, order_item.working_indicator);
+  log::debug("status={}"_sv, status);
+  auto side = json::map(order_item.side);
+  auto external_account = fmt::format("{}"_sv, order_item.account);  // XXX alloc
+  auto external_order_id = order_item.order_id;
+  auto order_type = json::map(order_item.ord_type);
+  auto time_in_force = json::map(order_item.time_in_force);
+  auto request_status = compute_request_status(order_item.ord_status);
+  auto error = json::guess_error(order_item.ord_rej_reason);
+  auto request_id =
+      request_type != RequestType::CANCEL_ORDER ? order_item.cl_ord_id : std::string_view{};
+  server::OMS_Ack ack{
+      .type = request_type,
+      .origin = Origin::EXCHANGE,
+      .status = request_status,
+      .error = error,
+      .text = order_item.text,
+      .version = version,
+      .request_id = request_id,
+      .quantity = order_item.order_qty,
+      .price = order_item.price,
+  };
+  roq::OrderUpdate order_update{
+      .stream_id = stream_id_,
+      .account = account_,
+      .order_id = order_id,
+      .exchange = Flags::exchange(),
+      .symbol = order_item.symbol,
+      .side = side,
+      .position_effect = {},
+      .max_show_quantity = NaN,
+      .order_type = order_type,
+      .time_in_force = time_in_force,
+      .execution_instruction = {},
+      .order_template = {},
+      .create_time_utc = {},
+      .update_time_utc = order_item.timestamp,  // XXX transact_time?
+      .external_account = external_account,
+      .external_order_id = external_order_id,
+      .status = status,
+      .quantity = order_item.order_qty,
+      .price = order_item.price,
+      .stop_price = order_item.stop_px,
+      .remaining_quantity = order_item.leaves_qty,
+      .traded_quantity = order_item.cum_qty,
+      .average_traded_price = order_item.avg_px,
+      .last_traded_quantity = NaN,
+      .last_traded_price = NaN,
+      .last_liquidity = {},
+      .routing_id = {},
+      .max_request_version = {},
+      .max_response_version = {},
+      .max_accepted_version = {},
+  };
+  if (shared_.update_order(
+          order_item.cl_ord_id,
+          stream_id_,
+          trace_info,
+          ack,
+          order_update,
+          []([[maybe_unused]] auto &order) {})) {
+  } else {
+    log::warn("*** EXTERNAL ORDER ***"_sv);
+    log::warn("order_item={}"_sv, order_item);
+  }
+}
+
+void OrderUpdate::operator()(
+    const json::Order &order,
+    const server::TraceInfo &trace_info,
+    RequestType request_type,
+    uint8_t user_id,
+    uint32_t order_id,
+    uint32_t version) {
+  log::debug("order={}"_sv, order);
+  for (auto &iter : order.data)
+    (*this)(iter, trace_info, request_type, user_id, order_id, version);
 }
 
 }  // namespace bitmex